Sorts forms by updated on Forms page (#3378)
* Updates getAllTemplates to optionally sort by updated_at

* Refactors DB call from component to Page

* Move MenuDropdown component and delete old myforms clientComponents

* update import

* Revalidate forms path on update/create form

* Simplify date formatting, use ISO8601 format

* Move pages components up a level no need for ...path

* Update imports

* Use status for searchParam filter

* Improved suspense/skeleton

---------
