+++ app/(gcforms)/[locale]/(form administration)/form-builder/actions.ts
@@ -47,6 +47,8 @@ export const createOrUpdateTemplate = async ({
   deliveryOption,
   securityAttribute,
 }: CreateOrUpdateTemplateType) => {
+  revalidatePath("/[locale]/forms", "page");
+
   if (id) {
     return updateTemplate({ id, formConfig, name, deliveryOption, securityAttribute });
   }
