+++ app/(gcforms)/[locale]/(form administration)/forms/components/client/ConfirmDelete.tsx
