+++ app/(gcforms)/[locale]/(form administration)/forms/components/client/MenuDropdown/MenuDropdown.test.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { render, cleanup, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import { MenuDropdown } from "@clientComponents/myforms/MenuDropdown/MenuDropdown";
+import { MenuDropdown } from "./MenuDropdown";
 
 const menuData = {
   id: "1",
